Undo GameLauncher changes
