+++ src/main/justacommonguy/battleshipgui/Ship.java
@@ -3,6 +3,7 @@ package justacommonguy.battleshipgui;
 import java.util.ArrayList;
 
 import justacommonguy.battleshipgui.gui.AllyCell;
+import justacommonguy.battleshipgui.gui.BattleshipGUI;
 
 public class Ship {
 
@@ -58,7 +59,7 @@ public class Ship {
 				return null;
 			}
 
-			AllyCell targetCell = GameLauncher.gameGUI.player.getCell(updatedLocation);
+			AllyCell targetCell = BattleshipGUI.gameGUI.player.getCell(updatedLocation);
 			if (targetCell != null) {
 				//// System.out.println("New location of " + this + ": " + updatedLocation);
 				newLocations.add(updatedLocation);
